Try to compute similarity in same job that runs netstruct

+++ utils/scripts/collect_tree_size_to_csv.py
@@ -16,7 +16,7 @@ from utils.loader import Timer
 def collect_tree_sizes_per_class(paths_helper, class_name, window_size, df):
     hash_json = paths_helper.hash_windows_list_template.format(class_name=class_name)
     raw_dict = load_dict_from_json(hash_json)
-    length_dict = {k: float(len(v) * window_size) for (k, v) in raw_dict.items()}
+    length_dict = {k: int(len(v) * window_size) for (k, v) in raw_dict.items()}
     length_dict_path = paths_helper.hash_winds_lengths_template.format(class_name=class_name)
     with open(length_dict_path, "w") as f:
         json.dump(length_dict, f)
